+++ Router/route_manager.py
@@ -6,6 +6,8 @@ from pathlib import Path
 from time import time, sleep
 from threading import Thread
 
+from sqlalchemy import case
+
 from config import config # type: ignore
 from utils.debug import Debug, catch_exceptions
 
@@ -14,9 +16,10 @@ from .context import Context
 from .ship import Ship
 from .route import Route
 
-SAVE_VARS:dict = {'system': '', 'src': '', 'dest': '', 'last_plot':
-                  'Neutron', 'neutron_params': {}, 'galaxy_params': {},
-                  'ship_id': '', 'cargo': 0, 'used_ships': [], 'history': []}
+SAVE_VARS:dict = {'system': '', 'src': '', 'dest': '', 'last_plot': 'Neutron',
+                  'carrier_id': '', 'carrier_state': 'Idle',
+                  'neutron_params': {}, 'galaxy_params': {},
+                  'ship_id': '', 'cargo': 0, 'shiplist': [], 'history': []}
 class Router():
     """
     Class to manage routes, all the route data and state information.
@@ -45,7 +48,7 @@ class Router():
         self.ship:Ship|None = None
 
         # Record of used ships and shipyard
-        self.used_ships:list = []
+        self.shiplist:list = []
         self.ships:dict[str, Ship] = {}
         self.history:list = []
 
@@ -53,8 +56,12 @@ class Router():
         self.last_plot:str = "Neutron"
         self.galaxy_params:dict = {}
         self.neutron_params:dict = {}
-
         self.cancel_plot:bool = False
+
+        # Carrier
+        self.carrier_id:str = ''
+        self.carrier_state:str = 'Idle'
+
         self._load()
         self._initialized = True
 
@@ -67,7 +74,7 @@ class Router():
         # Normalize ship_id to string since stored keys are strings
         sid = str(ship_id)
         if sid not in self.ships.keys():
-            Debug.logger.debug(f"ShipID {sid} not found in shipyard")
+            Debug.logger.info(f"ShipID {sid} not found in shipyard")
             self.ship_id = ""
             self.ship = None
             return
@@ -88,13 +95,17 @@ class Router():
         self.neutron_params['supercharge_mult'] = ship.supercharge_mult
         self.neutron_params['range'] = ship.range
         self.ships[self.ship_id] = ship
+
+        if self.ship_id in self.shiplist:
+            self.shiplist.remove(self.ship_id)
+        self.shiplist.insert(0, self.ship_id)
         Context.ui.switch_ship(self.ship)
 
 
     def jumped(self, system:str, entry:dict) -> None:
-        """ Called after a jump in order to update the route, the UI etc."""
+        """ Called after a carrier jump in order to update the route, the UI etc."""
 
-        Debug.logger.debug(f"Jumped called")
+        if Context.route.route == [] or Context.route.fleetcarrer == True: return
 
         Context.router.system = entry.get('StarSystem', system)
         Context.route.record_jump(entry.get('StarSystem', system), entry.get('JumpDist', 0))
@@ -110,6 +121,43 @@ class Router():
             Context.ui.update_waypoint()
 
 
+    def carrier_event(self, entry:dict) -> None:
+        """ Note carrier jumps for a cooldown notification """
+        if Context.route.route == [] or Context.route.fleetcarrer == False: return
+
+        match entry.get('event'):
+            case 'CarrierJumpRequest' if entry.get('SystemName', '') == Context.route.next_stop():
+                self.carrier_id = entry.get('CarrierID', '')
+                self.carrier_state = 'Jumping'
+                Debug.logger.debug(f"Carrier {self.carrier_id} jumping to {entry.get('SystemName', '')}")
+
+            case 'CarrierJumpCancelled' if self.carrier_id == entry.get('CarrierID', ''):
+                self.carrier_state = 'Cooldown'
+                Context.ui.parent.after(300000, lambda: self.cooldown_complete())
+
+            case 'CarrierLocation' if self.carrier_state == 'Jumping' and self.carrier_id == entry.get('CarrierID', '') and Context.ui.parent != None:
+                system:str = entry.get('StarSystem', '')
+                Debug.logger.debug(f"Carrier is in {system}")
+                if Context.route.update_route(0, system) < 0: return
+
+                Debug.logger.debug(f"Updated route")
+                self.carrier_state = 'Cooldown'
+                self.system = system
+                Context.ui.parent.after(300000, lambda: self.cooldown_complete())
+                Context.ui.update_waypoint()
+
+            #case _ if self.carrier_id == entry.get('CarrierID', ''):
+            #    self.carrier_state = 'Idle'
+
+
+    def cooldown_complete(self) -> None:
+        """ Show an informational messagebox indicating a carrier cooldown has completed. """
+        Debug.logger.debug(f"Cooldown complete notification triggered.")
+        if Context.ui.parent == None: return
+        self.carrier_state = 'Idle'
+        Context.ui.cooldown_complete()
+
+
     def _store_history(self) -> None:
         """ Upon route completion store src, dest and ship data """
 
@@ -118,14 +166,6 @@ class Router():
         if self.dest != '' and self.dest not in self.history:
             self.history.insert(0, Context.route.destination())
         self.history = list(dict.fromkeys(self.history))[:10] # Keep only last 10 unique entries
-
-        if self.ship_id == None or self.ship == None:
-            Debug.logger.debug(f"No ship to store")
-            return
-
-        if self.ship_id not in self.used_ships:
-            self.used_ships.append(self.ship_id)
-        Debug.logger.debug(f"Storing ship {str(self.ship)}")
 
 
     def plot_route(self, which:str, params:dict) -> bool:
@@ -143,7 +183,7 @@ class Router():
                 self.dest = params['to']
                 self.neutron_params = params
             case _:
-                Debug.logger.debug(f"Unknown route type {which}")
+                Debug.logger.error(f"Unknown route type {which}")
                 return False
 
         self.last_plot = which
@@ -159,8 +199,9 @@ class Router():
         self.cancel_plot = False
         try:
             limit:int = int(params.get('max_time', 20))
-            results:Response = requests.post(url, data=params, headers={'User-Agent': Context.plugin_useragent,
-                                                                        'Content-Type': 'application/x-www-form-urlencoded; charset=UTF-8'})
+            results:Response = requests.post(url, data=params,
+                                             headers={'User-Agent': Context.plugin_useragent,
+                                                      'Content-Type': 'application/x-www-form-urlencoded; charset=UTF-8'})
 
             if results.status_code != 202:
                 self.plot_error(results)
@@ -186,9 +227,7 @@ class Router():
             result:dict = json.loads(route_response.content)["result"]
             res:list = result.get('jumps', result.get('system_jumps', []))
 
-            cols:list = []
-            hdrs:list = []
-            h:str
+            cols:list = []; hdrs:list = []; h:str
             for h in HEADERS:
                 k:str
                 for k in res[0].keys():
@@ -271,7 +310,7 @@ class Router():
         """ Save a route to a CSV file """
         try:
             if Context.csv == None or Context.csv.write(Context.route.hdrs, Context.route.route) == False:
-                Debug.logger.debug(f"Failed to save route")
+                Debug.logger.error(f"Failed to save route")
                 Context.ui.show_error(errs['no_filename'])
                 return False
             return True
@@ -378,5 +417,9 @@ class Router():
         Context.route = Route(hdrs, route, offset, jumps)
         self.ship = Ship(dict.get('ship', {}))
         self.ships = {k: Ship(data) for k, data in dict.get('ships', {}).items()}
+
+        # Migrate
+        if self.shiplist == [] and self.ships != {}:
+            self.shiplist = [id for id in self.ships.keys()]