+++ load.py
@@ -1,10 +1,13 @@
 import tkinter as tk
+from tkinter import ttk
+import myNotebook as nb # type: ignore
+
 from pathlib import Path
 from semantic_version import Version #type: ignore
 
 from config import appname  # type: ignore
 
-from Router.constants import GIT_PROJECT, NAME, errs
+from Router.constants import GH_PROJECT, NAME, errs
 from utils.debug import Debug, catch_exceptions
 from utils.updater import Updater
 
@@ -13,9 +16,7 @@ from Router.route_manager import Router
 from Router.csv import CSV
 from Router.ui import UI
 
-
 def plugin_start3(plugin_dir: str) -> str:
-    # Debug Class
     Debug(plugin_dir)
 
     Context.plugin_name = NAME
@@ -26,9 +27,10 @@ def plugin_start3(plugin_dir: str) -> str:
     if version_file.is_file():
         version = Version(version_file.read_text())
     Context.plugin_version = version
-    Context.plugin_useragent = f"{GIT_PROJECT}-{version}"
+    Context.plugin_useragent = f"{GH_PROJECT}-{version}"
     Context.updater = Updater(str(Context.plugin_dir))
     Context.updater.check_for_update(Context.plugin_version)
+
     return NAME
 
 
@@ -43,27 +45,28 @@ def plugin_stop() -> None:
         Context.updater.install()
 
 
+def plugin_app(parent:tk.Widget) -> tk.Frame:
+    Context.csv = CSV()
+    Context.router = Router()
+    Context.ui = UI(parent)
+
+    return Context.ui.frame
+
+
 def journal_entry(cmdr:str, is_beta:bool, system:str, station:str, entry:dict, state:dict) -> None:
     match entry['event']:
         case 'Startup':
             Context.router.system = system
         case 'FSDJump' | 'Location' | 'SupercruiseExit' if entry.get('StarSystem', system) != Context.router.system:
             Context.router.jumped(system, entry)
-        case 'CarrierJump' if entry.get('Docked', True) == True:
-            Context.router.jumped(system, entry)
+        case 'CarrierJumpRequest' | 'CarrierLocation' | 'CarrierJumpCancelled':
+            Context.router.carrier_event(entry)
         case 'Loadout':
             Context.router.set_ship(entry)
         case 'ShipyardSwap':
             Context.router.swap_ship(entry.get('ShipID', ''))
         case 'Cargo':
             Context.router.cargo = entry.get('Count', 0)
-
-
-def plugin_app(parent:tk.Widget) -> tk.Frame:
-    Context.csv = CSV()
-    Context.router = Router()
-    Context.ui = UI(parent)
-    return Context.ui.frame
 
 
 def __version__() -> str: