+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # EDMC-NeutronDancer Changelog
 
+## v1.6.0 2026-01-##
+
+### New Features
+* Improved fleetcarrier support with jump cooldown notification
+* More detailed help
+* Added a refuel reminder icon
+
+### Bug Fixes
+* Fixed error with displaying booleans on imported routes
+* Yet another go at Linux copy to clipboard
+* Made ship menus update dynamically
+
+## v1.5.3 - 2026-01-03
+
+### Bug Fixes
 ## v1.5.3 - 2026-01-03
 
 ### Bug fixes
@@ -14,7 +29,7 @@
 
 ## v1.5.1 - 2025-12-31
 
-### Bug fixes
+### Bug Fixes
 * Fixed galaxy router error in dark mode when ships hadn't been switched
 * Addressed more dark and transparent mode craziness
 
@@ -67,12 +82,12 @@
 ### New Features
 * Added CSV importing of various formats
 
-### Bug fixes
+### Bug Fixes
 * Fixed Linux copy to clipboard
 
 ## v0.3.1 - 2025-12-14
 
-### Bug fixes
+### Bug Fixes
 * Fixed the updater and version identification
 
 
@@ -90,7 +105,7 @@
 ### Changes
 * Work on the autoupdater
 
-### Bug fixes
+### Bug Fixes
 * Autocreation of data directory
 * Clearing of placeholder text on click